+++ src/Router.js
@@ -3,6 +3,7 @@ import { Scene, Router, Actions } from 'react-native-router-flux';
 import LoginForm from './components/LoginForm';
 import EmployeeList from './components/EmployeeList';
 import EmployeeCreate from './components/EmployeeCreate';
+import EmployeeEdit from './components/EmployeeEdit';
 
 const RouterComponent = () => {
     return (
@@ -22,6 +23,7 @@ const RouterComponent = () => {
                         initial
                     />
                     <Scene key="employeeCreate" sceneStyle={styles.formStyle} component={EmployeeCreate} title="Create Employee" />
+                    <Scene key="employeeEdit" sceneStyle={styles.formStyle} component={EmployeeEdit} title="Edit Employee" />
                 </Scene>
             </Scene>
         </Router>
