+++ src/components/EmployeeCreate.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import { Picker, Text } from 'react-native';
 import { connect } from 'react-redux';
 import { employeeUpdate, employeeCreate } from "../actions/EmployeeActions";
-import { Card, CardSection, Input, Button } from './common';
+import { Card, CardSection, Button } from './common';
+import EmployeeForm from './EmployeeForm';
 
 class EmployeeCreate extends Component {
     onButtonPress() {
@@ -14,38 +14,7 @@ class EmployeeCreate extends Component {
     render() {
         return (
             <Card>
-                <CardSection>
-                    <Input
-                        label="Name"
-                        placeholder="Jane"
-                        value={this.props.name}
-                        onChangeText={text => this.props.employeeUpdate({prop: 'name', value: text})}
-                    />
-                </CardSection>
-                <CardSection>
-                    <Input
-                        label="Phone"
-                        placeholder="555-555-5555"
-                        value={this.props.phone}
-                        onChangeText={text => this.props.employeeUpdate({prop: 'phone', value: text})}
-                    />
-                </CardSection>
-                <CardSection style={{}}>
-                    <Text style={styles.pickerTextStyle}>Shift</Text>
-                    <Picker
-                        style={{flex: 1 }}
-                        selectedValue={this.props.shift}
-                        onValueChange={text => this.props.employeeUpdate({prop: 'shift', value: text})}
-                    >
-                        <Picker.Item label="Monday" value="Monday" />
-                        <Picker.Item label="Tuesday" value="Tuesday" />
-                        <Picker.Item label="Wednesday" value="Wednesday" />
-                        <Picker.Item label="Thursday" value="Thursday" />
-                        <Picker.Item label="Friday" value="Friday" />
-                        <Picker.Item label="Saturday" value="Saturday" />
-                        <Picker.Item label="Sunday" value="Sunday" />
-                    </Picker>
-                </CardSection>
+                <EmployeeForm {...this.props} />
                 <CardSection>
                     <Button onPress={this.onButtonPress.bind(this)}>
                         Create
@@ -56,12 +25,7 @@ class EmployeeCreate extends Component {
     }
 }
 
-const styles = {
-    pickerTextStyle: {
-        fontSize: 18,
-        paddingLeft: 20
-    }
-}
+
 
 const mapStateToProps = (state) => {
     const { name, phone, shift } = state.employeeForm;
