+++ src/components/EmployeeEdit.js
@@ -1,0 +1,39 @@
+import _ from 'lodash';
+import React, {Component} from 'react';
+import {connect} from 'react-redux';
+import {Card, CardSection, Button} from './common';
+import {employeeUpdate} from '../actions';
+import EmployeeForm from './EmployeeForm';
+
+class EmployeeEdit extends Component {
+  componentWillMount() {
+    _.each(this.props.employee, (value, prop) => {
+      this.props.employeeUpdate({prop, value});
+    });
+  }
+
+  onButtonPress() {
+    const {name, phone, shift} = this.props;
+  }
+
+
+  render() {
+    return (
+      <Card>
+        <EmployeeForm/>
+        <CardSection>
+          <Button onPress={this.onButtonPress.bind(this)}>
+            Save Changes
+          </Button>
+        </CardSection>
+      </Card>
+    )
+  }
+}
+
+const mapStateToProps = (state) => {
+  const {name, phone, shift} = state.employeeForm;
+  return {name, phone, shift};
+};
+
+export default connect(null, {employeeUpdate})(EmployeeEdit);
